show "Open with Database Tools" action in both IU and IC.
