+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-database/src/main/java/com/microsoft/azure/toolkit/intellij/database/mysql/IntellijMySqlActionsContributor.java
@@ -5,27 +5,40 @@
 
 package com.microsoft.azure.toolkit.intellij.database.mysql;
 
+import com.intellij.ide.plugins.PluginManagerCore;
 import com.intellij.openapi.actionSystem.AnActionEvent;
+import com.intellij.openapi.extensions.PluginId;
+import com.intellij.openapi.project.Project;
 import com.microsoft.azure.toolkit.ide.common.IActionsContributor;
 import com.microsoft.azure.toolkit.ide.common.action.ResourceCommonActionsContributor;
 import com.microsoft.azure.toolkit.ide.database.mysql.MySqlActionsContributor;
+import com.microsoft.azure.toolkit.intellij.common.action.IntellijActionsContributor;
 import com.microsoft.azure.toolkit.intellij.connector.ConnectorDialog;
+import com.microsoft.azure.toolkit.intellij.database.OpenInDatabaseToolsAction;
 import com.microsoft.azure.toolkit.intellij.database.connection.SqlDatabaseResource;
 import com.microsoft.azure.toolkit.intellij.database.mysql.connection.MySqlDatabaseResourceDefinition;
 import com.microsoft.azure.toolkit.intellij.database.mysql.creation.CreateMySqlAction;
 import com.microsoft.azure.toolkit.intellij.database.mysql.creation.MySqlCreationDialog;
 import com.microsoft.azure.toolkit.lib.common.action.AzureActionManager;
+import com.microsoft.azure.toolkit.lib.common.exception.AzureToolkitRuntimeException;
 import com.microsoft.azure.toolkit.lib.common.model.AzResource;
+import com.microsoft.azure.toolkit.lib.common.operation.AzureOperation;
 import com.microsoft.azure.toolkit.lib.common.task.AzureTaskManager;
 import com.microsoft.azure.toolkit.lib.database.DatabaseServerConfig;
+import com.microsoft.azure.toolkit.lib.database.JdbcUrl;
 import com.microsoft.azure.toolkit.lib.mysql.AzureMySql;
 import com.microsoft.azure.toolkit.lib.mysql.MySqlServer;
 import com.microsoft.azure.toolkit.lib.resource.ResourceGroup;
 
+import javax.annotation.Nonnull;
+import java.util.Objects;
 import java.util.function.BiConsumer;
 import java.util.function.BiPredicate;
 
 public class IntellijMySqlActionsContributor implements IActionsContributor {
+    private static final String NAME_PREFIX = "MySQL - %s";
+    private static final String DEFAULT_DRIVER_CLASS_NAME = "com.mysql.cj.jdbc.Driver";
+
     @Override
     public void registerHandlers(AzureActionManager am) {
         final BiPredicate<Object, AnActionEvent> condition = (r, e) -> r instanceof AzureMySql;
@@ -49,6 +62,26 @@ public class IntellijMySqlActionsContributor implements IActionsContributor {
             CreateMySqlAction.create(e.getProject(), config);
         };
         am.registerHandler(MySqlActionsContributor.GROUP_CREATE_MYSQL, (r, e) -> true, groupCreateMySqlHandler);
+
+        final BiConsumer<AzResource<?, ?, ?>, AnActionEvent> openDatabaseHandler = (c, e) -> openDatabaseTool(e.getProject(), (MySqlServer) c);
+        am.registerHandler(MySqlActionsContributor.OPEN_DATABASE_TOOL, (r, e) -> true, openDatabaseHandler);
+    }
+
+    @AzureOperation(name = "mysql.open_database_tools.server", params = {"server.getName()"}, type = AzureOperation.Type.ACTION)
+    private void openDatabaseTool(Project project, @Nonnull MySqlServer server) {
+        final String DATABASE_TOOLS_PLUGIN_ID = "com.intellij.database";
+        final String DATABASE_PLUGIN_NOT_INSTALLED = "\"Database tools and SQL\" plugin is not installed.";
+        final String NOT_SUPPORT_ERROR_ACTION = "\"Database tools and SQL\" plugin is only provided in IntelliJ Ultimate edition.";
+        if (PluginManagerCore.getPlugin(PluginId.findId(DATABASE_TOOLS_PLUGIN_ID)) == null) {
+            throw new AzureToolkitRuntimeException(DATABASE_PLUGIN_NOT_INSTALLED, NOT_SUPPORT_ERROR_ACTION, IntellijActionsContributor.TRY_ULTIMATE);
+        }
+        final OpenInDatabaseToolsAction.DatasourceProperties properties = OpenInDatabaseToolsAction.DatasourceProperties.builder()
+            .name(String.format(NAME_PREFIX, server.getName()))
+            .driverClassName(DEFAULT_DRIVER_CLASS_NAME)
+            .url(JdbcUrl.mysql(Objects.requireNonNull(server.getFullyQualifiedDomainName())).toString())
+            .username(server.getAdminName() + "@" + server.getName())
+            .build();
+        AzureTaskManager.getInstance().runLater(() -> OpenInDatabaseToolsAction.openDataSourceManagerDialog(project, properties));
     }
 
     @Override
