--- PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-database/src/main/resources/META-INF/azure-intellij-plugin-database-dbtools.xml
@@ -1,7 +1,0 @@
-<idea-plugin>
-    <extensions defaultExtensionNs="com.microsoft.tooling.msservices.intellij.azure">
-        <actions implementation="com.microsoft.azure.toolkit.intellij.database.mysql.IntellijMySqlActionsContributorForUltimate"/>
-        <actions implementation="com.microsoft.azure.toolkit.intellij.database.postgre.IntellijPostgreSqlActionsContributorForUltimate"/>
-        <actions implementation="com.microsoft.azure.toolkit.intellij.database.sqlserver.IntellijSqlServerActionsContributorForUltimate"/>
-    </extensions>
-</idea-plugin>
